+++ tests/stubs/enums.php
@@ -1,0 +1,7 @@
+<?php // phpcs:disable
+
+enum ThingStatus: string
+{
+    case FOO = 'foo';
+    case BAR = 'bar';
+}